Rename to 'draw' because it's a function. A function should be a verb.
